+++ apps/nft-marketplace-premium/src/modules/wizard/components/sections/NftDropSection.tsx
@@ -25,7 +25,6 @@ import {
   useOwnedNFTs,
   useUnclaimedNFTSupply,
 } from '@thirdweb-dev/react';
-import { ContractWrapper } from '@thirdweb-dev/sdk/dist/declarations/src/evm/core/classes/contract-wrapper';
 import { useWeb3React } from '@web3-react/core';
 import { BigNumber, utils } from 'ethers';
 import { useMemo, useState } from 'react';
@@ -128,8 +127,7 @@ export default function NftDropSection({ section }: NftDropSectionProps) {
 
   const isOpenEdition = useMemo(() => {
     if (contract) {
-      const contractWrapper = (contract as any)
-        .contractWrapper as ContractWrapper<any>;
+      const contractWrapper = (contract as any).contractWrapper as any;
 
       const featureDetected = detectContractFeature(
         contractWrapper,