+++ packages/dexkit-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dexkit/core
 
+## 0.14.0
+
+### Minor Changes
+
+- 7ee615d: Add unlock widget
+- 7ee615d: Add aidrop claimable erc20 feature. This allows users to reward communities without spending gas, and it was added as well option to add merkle tree.
+
 ## 0.13.0
 
 ### Minor Changes